+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * configure pyrefly
 * added pyright support
+* update deprecation handling to new django versioning scheme
 
 ## 3.1.0